move kinesis stream to new folder; add Write() method for logstream
